+++ server/game.js
@@ -1,48 +1,151 @@
 import { Actions } from "./protocol.js";
+import PokePicker from "./picker.js";
 
 class Player {
-  constructor(socket) {
+  constructor(name, socket) {
+    this.name = name; 
     this.socket = socket; 
     this.closed = false; 
     this.socket.on('close', () => this.closed = true); 
   }
+}
 
-  close(code, reason) {
-    if(!this.closed) {
-      this.socket.close(code, reason); 
-      this.closed = true; 
-    }
+class Round {
+  constructor(choices, answer, players) {
+    this.choices = choices; 
+    this.answer = answer; 
+    this.players = players; 
+  }
+
+  run(timeout) {
+    return new Promise((resolve) => {
+      const results = new Array(this.players.length).fill('');  
+      const callbacks = new Array(this.players.length); 
+      const cleanup = () => {
+        for(let i = 0; i < this.players.length; ++i) {
+          const { socket } = this.players[i]; 
+          socket.removeEventListener('message', callbacks[i].onmessage); 
+          socket.removeEventListener('message', callbacks[i].onclose); 
+        }
+        for(let i = 0; i < this.players.length; ++i) {
+          const { socket } = this.players[i]; 
+          socket.send(JSON.stringify({
+            action: Actions.ANSWER,
+            answer: this.answer,
+            correct: this.choices[this.answer] === results[i] 
+          })); 
+        } 
+        resolve(results); 
+      }
+      for(const { socket } of this.players) {
+        socket.send(JSON.stringify({
+          action: Actions.QUESTION,
+          choices: this.choices
+        }));
+      }
+      let responses = 0; 
+      for(let i = 0; i < this.players.length; ++i) {
+        const { name, socket } = this.players[i]; 
+        callbacks[i] = {
+          onmessage: ({ data }) => { 
+            try {
+              const { action, selection } = JSON.parse(data);              
+              if(action === Actions.RESPOND) {
+                if(!selection) {
+                  throw new Error('No selection specified'); 
+                }
+                if(!this.choices[selection]) {
+                  throw new Error('Invalid selection'); 
+                }
+                if(results[i]) {
+                  throw new Error('Already responded'); 
+                }
+                results[i] = this.choices[selection]; 
+                responses++; 
+                this.players.forEach(({ socket }) => {
+                  socket.send(JSON.stringify({
+                    action: Actions.RESPONDED,
+                    name
+                  })); 
+                }); 
+                if(responses >= this.players.length) {
+                  cleanup(); 
+                }
+              }
+            } catch({ message }) {
+              socket.send(JSON.stringify({
+                action: Actions.ERROR,
+                message 
+              }));  
+            }
+          },
+          onclose: () => {
+            if(!results[i]) {
+              responses++; 
+              if(responses >= this.players.length) {
+                cleanup(); 
+              }
+            }
+          }
+        }; 
+        socket.addEventListener('message', callbacks[i].onmessage); 
+        socket.addEventListener('close', callbacks[i].onclose); 
+      }
+      if(timeout != Infinity) {
+        setTimeout(cleanup, timeout); 
+      }
+    }); 
   }
 }
 
 class Game {
-  constructor(game_reg, id, hostname, hostws) {
+  constructor(game_reg, id, hostname, hostws, options = {}) {
+    const {
+      gen = 0,
+      rounds = 10,
+      count = 4,
+      timeout = Infinity // ie no timeout
+    } = options; 
     this.game_reg = game_reg; 
     this.id = id; 
     this.started = false; 
     this.players = new Map(); 
-    this.players.set(hostname, new Player(hostws));  
-    hostws.on('message', (msg) => {
+    this.players.set(hostname, new Player(hostname, hostws));  
+    this.gen = gen; 
+    this.rounds = rounds; 
+    this.count = count; 
+    this.timeout = timeout; 
+    hostws.on('message', async (msg) => {
       try {
-        const { action } = JSON.parse(msg); 
-        if(!action) {
-          throw new Error('No action specified'); 
-        }
+        const { action = Actions.NONE } = JSON.parse(msg); 
         switch(action) {
+          case Actions.NONE:
+            throw new Error('No action specified'); 
           case Actions.LEAVE:
-            this.remove_player(hostname); 
             if(!this.started) {
               this.cancel(); 
+            } else {
+              this.remove_player(hostname); 
             }
             break; 
           case Actions.START:
-            this.start(); 
+            if(!this.started) {
+              await this.start(); 
+            } else {
+              throw new Error(`Game '${this.id}' already started`); 
+            }
             break;
           case Actions.CANCEL:
-            this.cancel(); 
+            if(!this.started) {
+              this.cancel(); 
+            } else {
+              throw new Error(`Game '${this.id}' alreday started`); 
+            }
+            break; 
+          case Actions.RESPOND: 
             break; 
           default: 
-            throw new Error(`${action} action is not recognized`); 
+            throw new Error(`Action '${action}' not recognized`); 
         }
       } catch({ message }) {
         hostws.send(JSON.stringify({
@@ -52,9 +155,10 @@ class Game {
       }
     });
     hostws.on('close', () => {
-      this.remove_player(hostname); 
       if(!this.started) {
         this.cancel(); 
+      } else {
+        this.remove_player(hostname); 
       }
     }); 
   } 
@@ -63,18 +167,18 @@ class Game {
     if(this.players.has(name)) {
       ws.send(JSON.stringify({
         action: Actions.ERROR,
-        message: `Player ${name} already exists`
+        message: `Player '${name}' already exists`
       })); 
       return false; 
     }
     if(this.started) {
       ws.send(JSON.stringify({
         action: Actions.ERROR,
-        message: `Game ${this.id} has already started` 
+        message: `Game '${this.id}' already started` 
       }));
       return false; 
     }
-    this.players.set(name, new Player(ws)); 
+    this.players.set(name, new Player(name, ws)); 
     this.players.forEach(({ socket }) => { 
       socket.send(JSON.stringify({
         action: Actions.JOINED,
@@ -83,20 +187,21 @@ class Game {
     });
     ws.on('message', (msg) => {
       try {
-        const { action } = JSON.parse(msg); 
-        if(!action) {
-          throw new Error('No action specified'); 
-        }
+        const { action = Actions.NONE } = JSON.parse(msg); 
         switch(action) {
+          case Actions.NONE: 
+            throw new Error('No action specified'); 
           case Actions.LEAVE: 
             this.remove_player(name); 
             break;
           case Actions.START:
-            throw new Error('Only host can start game'); 
+            throw new Error(`Only host can start game '${this.id}'`); 
           case Actions.CANCEL:
-            throw new Error('Only host can cancel game'); 
+            throw new Error(`Only host can cancel game '${this.id}'`); 
+          case Actions.RESPOND: 
+            break; 
           default: 
-            throw new Error(`${action} action not recognized`); 
+            throw new Error(`Action '${action}' not recognized`);
         }
       } catch({ message }) {
         ws.send(JSON.stringify({
@@ -111,26 +216,38 @@ class Game {
 
   remove_player(name) {
     if(!this.players.has(name)) {
-      return false; 
+      return null; 
     }
-    this.players.get(name).close(); 
-    this.players.delete(name); 
+    const player = this.players.get(name); 
     this.players.forEach(({ socket }) => {
       socket.send(JSON.stringify({
         action: Actions.LEFT,
         name
       })); 
     });
-    return true;    
+    this.players.delete(name);
+    if(!player.closed) {
+      player.socket.removeAllListeners(); 
+      this.game_reg.wss.emit('connection', player.socket); 
+    } 
+    return player; 
   }
   
-  start() {
+  async start() {
     this.started = true; 
     for(const player of this.players.values()) {
       player.socket.send(JSON.stringify({
         action: Actions.STARTED
       })); 
     }
+    const picker = new PokePicker(); 
+    await picker.initialize(this.gen);  
+    for(let i = 0; i < this.rounds; ++i) {
+      const choices = picker.pick(this.count); 
+      const answer = Math.floor(Math.random() * this.count); 
+      const round = new Round(choices, answer, Array.from(this.players.values())); 
+      await round.run(this.timeout); 
+    } 
   }
 
   cancel() {
@@ -139,22 +256,23 @@ class Game {
         action: Actions.CANCELLED
       })); 
     }
-    for(const name of this.players.keys()) {
+    for(const { name } of this.players.values()) {
       this.remove_player(name); 
     }
-    this.game_reg.delete_game(this.id); 
+    this.game_reg.games.delete(this.id); 
   }
 }
 
 export default class GameRegistry {
   static ALPHANUM_CHARSET = 'abcdefghijklmnopqrstuvwxyz123456789'; 
 
-  constructor(options = {}) { 
+  constructor(wss, options = {}) { 
     const {
       charset = GameRegistry.ALPHANUM_CHARSET,
       id_len = 6,
       tries = Infinity
     } = options; 
+    this.wss = wss; 
     this.charset = charset; 
     this.id_len = Math.max(1, id_len); 
     this.tries = Math.max(1, tries); 
@@ -178,10 +296,10 @@ export default class GameRegistry {
     return null; 
   }
 
-  new_game(hostname, hostws) {
+  new_game(hostname, hostws, options = {}) {
     const id = this.gen_game_id(); 
     if(id) {
-      this.games.set(id, new Game(this, id, hostname, hostws)); 
+      this.games.set(id, new Game(this, id, hostname, hostws, options)); 
       hostws.send(JSON.stringify({
         action: Actions.HOSTED,
         id
@@ -194,9 +312,5 @@ export default class GameRegistry {
     }
     return id; 
   }
-  
-  delete_game(id) {
-    return this.games.delete(id); 
-  } 
 }